fast probe timer for NDP

+++ ndp/server/pktHandler.go
@@ -29,6 +29,7 @@ import (
 	"github.com/google/gopacket/layers"
 	"l3/ndp/config"
 	"l3/ndp/debug"
+	"net"
 	"reflect"
 	"utils/commonDefs"
 )
@@ -182,14 +183,16 @@ func (svr *NDPServer) CheckSrcMac(macAddr string) bool {
  */
 func (svr *NDPServer) insertNeigborInfo(nbrInfo *config.NeighborConfig) {
 	svr.NeigborEntryLock.Lock()
-	svr.NeighborInfo[nbrInfo.IpAddr] = *nbrInfo
-	svr.neighborKey = append(svr.neighborKey, nbrInfo.IpAddr)
+	nbrKey := createNeighborKey(nbrInfo.MacAddr, nbrInfo.IpAddr, nbrInfo.Intf)
+	svr.NeighborInfo[nbrKey] = *nbrInfo
+	svr.neighborKey = append(svr.neighborKey, nbrKey)
 	svr.NeigborEntryLock.Unlock()
 }
 
 func (svr *NDPServer) updateNeighborInfo(nbrInfo *config.NeighborConfig) {
 	svr.NeigborEntryLock.Lock()
-	svr.NeighborInfo[nbrInfo.IpAddr] = *nbrInfo
+	nbrKey := createNeighborKey(nbrInfo.MacAddr, nbrInfo.IpAddr, nbrInfo.Intf)
+	svr.NeighborInfo[nbrKey] = *nbrInfo
 	svr.NeigborEntryLock.Unlock()
 }
 
@@ -197,9 +200,9 @@ func (svr *NDPServer) updateNeighborInfo(nbrInfo *config.NeighborConfig) {
  *	deleteNeighborInfo: Helper API to update list of neighbor keys that are deleted by ndp
  *	@NOTE: caller is responsible for acquiring the lock to access slice
  */
-func (svr *NDPServer) deleteNeighborInfo(nbrIp string) {
+func (svr *NDPServer) deleteNeighborInfo(nbrKey string) {
 	for idx, _ := range svr.neighborKey {
-		if svr.neighborKey[idx] == nbrIp {
+		if svr.neighborKey[idx] == nbrKey {
 			svr.neighborKey = append(svr.neighborKey[:idx], svr.neighborKey[idx+1:]...)
 			break
 		}
@@ -215,40 +218,47 @@ func (svr *NDPServer) deleteNeighborInfo(nbrIp string) {
 func (svr *NDPServer) CreateNeighborInfo(nbrInfo *config.NeighborConfig) {
 	debug.Logger.Debug("Calling create ipv6 neighgor for global nbrinfo is", nbrInfo.IpAddr, nbrInfo.MacAddr,
 		nbrInfo.VlanId, nbrInfo.IfIndex)
-	_, err := svr.SwitchPlugin.CreateIPv6Neighbor(nbrInfo.IpAddr, nbrInfo.MacAddr,
-		nbrInfo.VlanId, nbrInfo.IfIndex)
-	if err != nil {
-		debug.Logger.Err("create ipv6 global neigbor failed for", nbrInfo, "error is", err)
-		// do not enter that neighbor in our neigbor map
-		return
+	if net.ParseIP(nbrInfo.IpAddr).IsLinkLocalUnicast() == false {
+		_, err := svr.SwitchPlugin.CreateIPv6Neighbor(nbrInfo.IpAddr, nbrInfo.MacAddr, nbrInfo.VlanId, nbrInfo.IfIndex)
+		if err != nil {
+			debug.Logger.Err("create ipv6 global neigbor failed for", nbrInfo, "error is", err)
+			// do not enter that neighbor in our neigbor map
+			return
+		}
 	}
 	svr.SendIPv6CreateNotification(nbrInfo.IpAddr, nbrInfo.IfIndex)
 	svr.insertNeigborInfo(nbrInfo)
 }
 
-func (svr *NDPServer) deleteNeighbor(nbrIp string, ifIndex int32) {
+func (svr *NDPServer) deleteNeighbor(nbrKey string, ifIndex int32) {
 	// Inform clients that neighbor is gonna be deleted
+	splitString := splitNeighborKey(nbrKey)
+	nbrIp := splitString[1]
 	svr.SendIPv6DeleteNotification(nbrIp, ifIndex)
+	//nbrInfo := svr.NeighborInfo[nbrKey]
 	// Request asicd to delete the neighbor
-	_, err := svr.SwitchPlugin.DeleteIPv6Neighbor(nbrIp)
-	if err != nil {
-		debug.Logger.Err("delete ipv6 neigbor failed for", nbrIp, "error is", err)
+	if net.ParseIP(nbrIp).IsLinkLocalUnicast() == false {
+		_, err := svr.SwitchPlugin.DeleteIPv6Neighbor(nbrIp)
+		if err != nil {
+			debug.Logger.Err("delete ipv6 neigbor failed for", nbrIp, "error is", err)
+		}
 	}
 	// delete the entry from neighbor map
-	delete(svr.NeighborInfo, nbrIp)
-	svr.deleteNeighborInfo(nbrIp)
+	delete(svr.NeighborInfo, nbrKey)
+	svr.deleteNeighborInfo(nbrKey)
 }
 
 func (svr *NDPServer) UpdateNeighborInfo(nbrInfo *config.NeighborConfig, oldNbrEntry config.NeighborConfig) {
 	//svr.SendIPv6DeleteNotification(oldNbrEntry.IpAddr, oldNbrEntry.IfIndex)
 	debug.Logger.Debug("Calling update ipv6 neighgor for global nbrinfo is", nbrInfo.IpAddr, nbrInfo.MacAddr,
 		nbrInfo.VlanId, nbrInfo.IfIndex)
-	_, err := svr.SwitchPlugin.UpdateIPv6Neighbor(nbrInfo.IpAddr, nbrInfo.MacAddr,
-		nbrInfo.VlanId, nbrInfo.IfIndex)
-	if err != nil {
-		debug.Logger.Err("update ipv6 global neigbor failed for", nbrInfo, "error is", err)
-		// do not enter that neighbor in our neigbor map
-		return
+	if net.ParseIP(nbrInfo.IpAddr).IsLinkLocalUnicast() == false {
+		_, err := svr.SwitchPlugin.UpdateIPv6Neighbor(nbrInfo.IpAddr, nbrInfo.MacAddr, nbrInfo.VlanId, nbrInfo.IfIndex)
+		if err != nil {
+			debug.Logger.Err("update ipv6 global neigbor failed for", nbrInfo, "error is", err)
+			// do not enter that neighbor in our neigbor map
+			return
+		}
 	}
 	//svr.SendIPv6CreateNotification(nbrInfo.IpAddr, nbrInfo.IfIndex)
 	svr.updateNeighborInfo(nbrInfo)
@@ -263,9 +273,9 @@ func (svr *NDPServer) UpdateNeighborInfo(nbrInfo *config.NeighborConfig, oldNbrE
  */
 func (svr *NDPServer) DeleteNeighborInfo(deleteEntries []string, ifIndex int32) {
 	svr.NeigborEntryLock.Lock()
-	for _, nbrIp := range deleteEntries {
+	for _, nbrKey := range deleteEntries {
 		//debug.Logger.Debug("Calling delete ipv6 neighbor for nbrIp:", nbrIp)
-		svr.deleteNeighbor(nbrIp, ifIndex)
+		svr.deleteNeighbor(nbrKey, ifIndex)
 	}
 	svr.NeigborEntryLock.Unlock()
 }
@@ -282,6 +292,7 @@ func (svr *NDPServer) ProcessRxPkt(ifIndex int32, pkt gopacket.Packet) error {
 	var l3Port Interface
 	var l2Port PhyPort
 	var exists bool
+	var nbrKey string
 	// if we receive packet on L2 Physical interface then the we need get l3 port via cross referencing PhyPortToL3PortMap
 	l3IfIndex, l3exists := svr.PhyPortToL3PortMap[ifIndex]
 	if l3exists {
@@ -318,12 +329,13 @@ func (svr *NDPServer) ProcessRxPkt(ifIndex int32, pkt gopacket.Packet) error {
 		nbrInfo.Intf = l2Port.Info.Name
 	}
 	svr.PopulateVlanInfo(nbrInfo, l3Port.IntfRef)
+	nbrKey = createNeighborKey(nbrInfo.MacAddr, nbrInfo.IpAddr, nbrInfo.Intf)
 	// based on operation program hardware, update sw & send notifications
 	switch operation {
 	case CREATE:
 		svr.CreateNeighborInfo(nbrInfo)
 	case UPDATE:
-		nbrEntry, exists := svr.NeighborInfo[nbrInfo.IpAddr]
+		nbrEntry, exists := svr.NeighborInfo[nbrKey]
 		if !exists { //entry does not exists and hence creating new
 			debug.Logger.Info("!!!!!!ALERT!!!!!! NDP Server does not have nbrInfo for ipaddr:",
 				nbrInfo.IpAddr, "hence on UPDATE doing CREATE")
@@ -338,7 +350,7 @@ func (svr *NDPServer) ProcessRxPkt(ifIndex int32, pkt gopacket.Packet) error {
 		}
 	case DELETE:
 		//svr.deleteNeighbor(nbrInfo.IpAddr, ifIndex) // used mostly by RA
-		svr.deleteNeighbor(nbrInfo.IpAddr, l3Port.IfIndex) // used mostly by RA
+		svr.deleteNeighbor(createNeighborKey(nbrInfo.MacAddr, nbrInfo.IpAddr, nbrInfo.Intf), l3Port.IfIndex) // used mostly by RA
 	}
 
 early_exit:
@@ -353,6 +365,7 @@ early_exit:
 func (svr *NDPServer) ProcessTimerExpiry(pktData config.PacketData) error {
 	var l3Port Interface
 	var exists bool
+	var intfName string
 	// if we receive packet on L2 Physical interface then the we need get l3 port via cross referencing PhyPortToL3PortMap
 	l3IfIndex, l3exists := svr.PhyPortToL3PortMap[pktData.IfIndex]
 	if l3exists {
@@ -361,26 +374,30 @@ func (svr *NDPServer) ProcessTimerExpiry(pktData config.PacketData) error {
 		if !exists {
 			return errors.New(fmt.Sprintln("Entry for ifIndex:", l3IfIndex, "doesn't exists"))
 		}
+		l2Port := svr.L2Port[pktData.IfIndex]
+		intfName = l2Port.Info.Name
 	} else {
 		// Port is the l3 port
 		l3Port, exists = svr.L3Port[pktData.IfIndex]
 		if !exists {
 			return errors.New(fmt.Sprintln("Entry for ifIndex:", pktData.IfIndex, "doesn't exists"))
 		}
+		intfName = l3Port.IntfRef
 	}
+	nbrKey := createNeighborKey(pktData.NeighborMac, pktData.NeighborIp, intfName)
 	// fix this when we have per port mac addresses
 	operation := l3Port.SendND(pktData, svr.SwitchMac)
 	if operation == DELETE {
 		//svr.deleteNeighbor(pktData.NeighborIp, pktData.IfIndex)
-		svr.deleteNeighbor(pktData.NeighborIp, l3Port.IfIndex)
+		svr.deleteNeighbor(nbrKey, l3Port.IfIndex)
 	}
 	if l3exists {
 		svr.L3Port[l3IfIndex] = l3Port
 	} else {
 		svr.L3Port[pktData.IfIndex] = l3Port
 	}
-	nbrInfo := svr.NeighborInfo[pktData.NeighborIp]
-	svr.NeighborInfo[pktData.NeighborIp] = nbrInfo
+	//nbrInfo := svr.NeighborInfo[nbrKey]
+	//svr.NeighborInfo[nbrKey] = nbrInfo
 	svr.counter.Send++
 	return nil
 }
